much cleaner as a mixin imo
